+++ lib/sprout/src/main/kotlin/org/partiql/sprout/generator/target/kotlin/KotlinGenerator.kt
@@ -1,5 +1,7 @@
 package org.partiql.sprout.generator.target.kotlin
 
+import com.squareup.kotlinpoet.ClassName
+import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.PropertySpec
@@ -9,7 +11,6 @@ import org.partiql.sprout.generator.Generator
 import org.partiql.sprout.generator.target.kotlin.poems.KotlinBuilderPoem
 import org.partiql.sprout.generator.target.kotlin.poems.KotlinJacksonPoem
 import org.partiql.sprout.generator.target.kotlin.poems.KotlinListenerPoem
-import org.partiql.sprout.generator.target.kotlin.poems.KotlinMetasPoem
 import org.partiql.sprout.generator.target.kotlin.poems.KotlinVisitorPoem
 import org.partiql.sprout.generator.target.kotlin.spec.KotlinFileSpec
 import org.partiql.sprout.generator.target.kotlin.spec.KotlinNodeSpec
@@ -35,9 +36,8 @@ class KotlinGenerator(private val options: KotlinOptions) : Generator<KotlinResu
                 "visitor" -> KotlinVisitorPoem(symbols)
                 "builder" -> KotlinBuilderPoem(symbols)
                 "listener" -> KotlinListenerPoem(symbols)
-                "metas" -> KotlinMetasPoem(symbols)
                 "jackson" -> KotlinJacksonPoem(symbols)
-                else -> error("unknown poem $it, expected: visitor, builder, listener, metas, jackson")
+                else -> error("unknown poem $it, expected: visitor, builder, listener, jackson")
             }
         }
 
@@ -62,8 +62,8 @@ class KotlinGenerator(private val options: KotlinOptions) : Generator<KotlinResu
     /**
      * Generate a NodeSpec for each Type in the given Universe
      */
-    private fun Universe.nodes(symbols: KotlinSymbols): List<KotlinNodeSpec> = types.mapNotNull { it.generate(symbols) }
-        .map {
+    private fun Universe.nodes(symbols: KotlinSymbols): List<KotlinNodeSpec> =
+        types.mapNotNull { it.generate(symbols) }.map {
             it.builder.superclass(symbols.base)
             it
         }
@@ -71,9 +71,8 @@ class KotlinGenerator(private val options: KotlinOptions) : Generator<KotlinResu
     /**
      * Generate all top-level enums as these are not children of a product definition
      */
-    private fun Universe.types(symbols: KotlinSymbols) = types.filterIsInstance<TypeDef.Enum>()
-        .map { it.generate(symbols) }
-        .toMutableList()
+    private fun Universe.types(symbols: KotlinSymbols) =
+        types.filterIsInstance<TypeDef.Enum>().map { it.generate(symbols) }.toMutableList()
 
     /**
      * Entry point for node generation.
@@ -87,25 +86,34 @@ class KotlinGenerator(private val options: KotlinOptions) : Generator<KotlinResu
     /**
      * Product Node Generation
      */
-    private fun TypeDef.Product.generate(symbols: KotlinSymbols) = KotlinNodeSpec.Product(
-        product = this,
-        props = props.map { KotlinNodeSpec.Prop(it.name.toCamelCase(), symbols.typeNameOf(it.ref)) },
-        nodes = children.mapNotNull { it.generate(symbols) },
-        clazz = symbols.clazz(ref),
-        ext = (props.enumProps(symbols) + types.enums(symbols)).toMutableList(),
-    ).apply {
-        props.forEach {
-            val para = ParameterSpec.builder(it.name, it.type).build()
-            val prop = PropertySpec.builder(it.name, it.type).initializer(it.name).build()
-            builder.addProperty(prop)
-            constructor.addParameter(para)
+    private fun TypeDef.Product.generate(symbols: KotlinSymbols): KotlinNodeSpec {
+        val clazz = symbols.clazz(ref)
+        val clazzImpl = ClassName(
+            packageName = clazz.packageName + ".impl",
+            simpleNames = listOf(symbols.pascal(ref) + "Impl"),
+        )
+        return KotlinNodeSpec.Product(
+            product = this,
+            props = props.map { KotlinNodeSpec.Prop(it.name.toCamelCase(), symbols.typeNameOf(it.ref)) },
+            implClazz = clazzImpl,
+            impl = TypeSpec.classBuilder(clazzImpl),
+            nodes = children.mapNotNull { it.generate(symbols) },
+            clazz = symbols.clazz(ref),
+            ext = (props.enumProps(symbols) + types.enums(symbols)).toMutableList(),
+        ).apply {
+            props.forEach {
+                val para = ParameterSpec.builder(it.name, it.type).build()
+                val prop = PropertySpec.builder(it.name, it.type).build()
+                builder.addProperty(prop.toBuilder().addModifiers(KModifier.ABSTRACT).build())
+                impl.addProperty(prop.toBuilder().addModifiers(KModifier.OVERRIDE).initializer(it.name).build())
+                constructor.addParameter(para)
+            }
+            // impls are open
+            impl.superclass(clazz)
+            nodes.forEach { it.builder.superclass(symbols.base) }
+
+            this.addDataClassMethods()
         }
-        when (options.node.modifier) {
-            KotlinNodeOptions.Modifier.FINAL -> {}
-            KotlinNodeOptions.Modifier.DATA -> if (props.isNotEmpty()) builder.addModifiers(KModifier.DATA)
-            KotlinNodeOptions.Modifier.OPEN -> builder.addModifiers(KModifier.OPEN)
-        }
-        nodes.forEach { it.builder.superclass(symbols.base) }
     }
 
     /**
@@ -125,9 +133,8 @@ class KotlinGenerator(private val options: KotlinOptions) : Generator<KotlinResu
     /**
      * Enum constant generation
      */
-    private fun TypeDef.Enum.generate(symbols: KotlinSymbols) = TypeSpec.enumBuilder(symbols.clazz(ref))
-        .apply { values.forEach { addEnumConstant(it) } }
-        .build()
+    private fun TypeDef.Enum.generate(symbols: KotlinSymbols) =
+        TypeSpec.enumBuilder(symbols.clazz(ref)).apply { values.forEach { addEnumConstant(it) } }.build()
 
     private fun List<TypeProp>.enumProps(symbols: KotlinSymbols) = filterIsInstance<TypeProp.Inline>().mapNotNull {
         when (it.def) {
@@ -138,5 +145,31 @@ class KotlinGenerator(private val options: KotlinOptions) : Generator<KotlinResu
 
     private fun List<TypeDef>.enums(symbols: KotlinSymbols) = filterIsInstance<TypeDef.Enum>().map {
         it.generate(symbols)
+    }
+
+    // TODO generate hashCode, equals, componentN so we can have OPEN internal implementations
+    private fun KotlinNodeSpec.Product.addDataClassMethods() {
+        impl.addModifiers(KModifier.INTERNAL, KModifier.OPEN)
+        // builder.addFunction(
+        //     FunSpec.builder("hashCode").addModifiers(KModifier.OVERRIDE, KModifier.ABSTRACT).returns(Int::class).build()
+        // )
+        // builder.addFunction(
+        //     FunSpec.builder("equals").addModifiers(KModifier.OVERRIDE, KModifier.ABSTRACT).returns(Boolean::class)
+        //         .addParameter(ParameterSpec.builder("other", Any::class.asTypeName().copy(nullable = true)).build())
+        //         .build()
+        // )
+        val copy = FunSpec.builder("copy").addModifiers(KModifier.ABSTRACT).returns(clazz)
+        val copyImpl = FunSpec.builder("copy")
+            .addModifiers(KModifier.OVERRIDE)
+            .returns(clazz)
+            .addStatement("return %T(${props.joinToString { it.name }})", implClazz)
+        props.forEach {
+            val para = ParameterSpec.builder(it.name, it.type).build()
+            copy.addParameter(para.toBuilder().defaultValue("this.${it.name}").build())
+            copyImpl.addParameter(para)
+        }
+
+        builder.addFunction(copy.build())
+        impl.addFunction(copyImpl.build())
     }
 }