+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/plan/RexConverter.kt
@@ -9,12 +9,9 @@ import org.partiql.lang.eval.builtins.ExprFunctionCurrentUser
 import org.partiql.lang.eval.err
 import org.partiql.lang.eval.errorContextFrom
 import org.partiql.lang.planner.transforms.AstToPlan
-import org.partiql.plan.Arg
-import org.partiql.plan.Branch
 import org.partiql.plan.Case
-import org.partiql.plan.Field
+import org.partiql.plan.Plan
 import org.partiql.plan.Rex
-import org.partiql.plan.Step
 import org.partiql.types.StaticType
 
 /**
@@ -57,11 +54,11 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     private fun convert(vararg nodes: PartiqlAst.Expr) = nodes.map { convert(it) }
 
     private fun arg(name: String, node: PartiqlAst.PartiqlAstNode) = when (node) {
-        is PartiqlAst.Expr -> Arg.Value(
+        is PartiqlAst.Expr -> Plan.argValue(
             name = name,
             value = convert(node),
         )
-        is PartiqlAst.Type -> Arg.Type(
+        is PartiqlAst.Type -> Plan.argType(
             name = name,
             type = TypeConverter.convert(node)
         )
@@ -100,12 +97,12 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     override fun walkMetas(node: MetaContainer, ctx: Ctx) = AstToPlan.unsupported(ctx.node)
 
     override fun walkExprMissing(node: PartiqlAst.Expr.Missing, ctx: Ctx) = visit(node) {
-        Rex.Lit(ionNull(), StaticType.MISSING)
+        Plan.rexLit(ionNull(), StaticType.MISSING)
     }
 
     override fun walkExprLit(node: PartiqlAst.Expr.Lit, ctx: Ctx) = visit(node) {
         val ionType = node.value.type.toIonType()
-        Rex.Lit(
+        Plan.rexLit(
             value = node.value,
             type = TypeConverter.convert(ionType)
         )
@@ -121,7 +118,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
                 internal = false
             )
         }
-        Rex.Call(
+        Plan.rexCall(
             id = functionName,
             args = emptyList(),
             type = null
@@ -129,7 +126,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprId(node: PartiqlAst.Expr.Id, ctx: Ctx) = visit(node) {
-        Rex.Id(
+        Plan.rexId(
             name = node.name.text,
             case = convertCase(node.case),
             qualifier = when (node.qualifier) {
@@ -141,16 +138,16 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprPath(node: PartiqlAst.Expr.Path, ctx: Ctx) = visit(node) {
-        Rex.Path(
+        Plan.rexPath(
             root = convert(node.root),
             steps = node.steps.map {
                 when (it) {
-                    is PartiqlAst.PathStep.PathExpr -> Step.Key(
+                    is PartiqlAst.PathStep.PathExpr -> Plan.stepKey(
                         value = convert(it.index),
                         case = convertCase(it.case)
                     )
-                    is PartiqlAst.PathStep.PathUnpivot -> Step.Unpivot()
-                    is PartiqlAst.PathStep.PathWildcard -> Step.Wildcard()
+                    is PartiqlAst.PathStep.PathUnpivot -> Plan.stepUnpivot()
+                    is PartiqlAst.PathStep.PathWildcard -> Plan.stepWildcard()
                 }
             },
             type = null,
@@ -158,7 +155,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprNot(node: PartiqlAst.Expr.Not, ctx: Ctx) = visit(node) {
-        Rex.Unary(
+        Plan.rexUnary(
             value = convert(node.expr),
             op = Rex.Unary.Op.NOT,
             type = StaticType.BOOL,
@@ -166,7 +163,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprPos(node: PartiqlAst.Expr.Pos, ctx: Ctx) = visit(node) {
-        Rex.Unary(
+        Plan.rexUnary(
             value = convert(node.expr),
             op = Rex.Unary.Op.POS,
             type = StaticType.NUMERIC,
@@ -174,7 +171,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprNeg(node: PartiqlAst.Expr.Neg, ctx: Ctx) = visit(node) {
-        Rex.Unary(
+        Plan.rexUnary(
             value = convert(node.expr),
             op = Rex.Unary.Op.NEG,
             type = StaticType.NUMERIC,
@@ -182,7 +179,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprPlus(node: PartiqlAst.Expr.Plus, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.PLUS,
@@ -191,7 +188,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprMinus(node: PartiqlAst.Expr.Minus, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.MINUS,
@@ -200,7 +197,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprTimes(node: PartiqlAst.Expr.Times, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.TIMES,
@@ -209,7 +206,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprDivide(node: PartiqlAst.Expr.Divide, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.DIV,
@@ -218,7 +215,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprModulo(node: PartiqlAst.Expr.Modulo, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.MODULO,
@@ -227,7 +224,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprConcat(node: PartiqlAst.Expr.Concat, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.CONCAT,
@@ -236,7 +233,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprAnd(node: PartiqlAst.Expr.And, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.AND,
@@ -245,7 +242,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprOr(node: PartiqlAst.Expr.Or, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.OR,
@@ -254,7 +251,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprEq(node: PartiqlAst.Expr.Eq, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.EQ,
@@ -263,7 +260,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprNe(node: PartiqlAst.Expr.Ne, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.NEQ,
@@ -272,7 +269,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprGt(node: PartiqlAst.Expr.Gt, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.GT,
@@ -281,7 +278,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprGte(node: PartiqlAst.Expr.Gte, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.GTE,
@@ -290,7 +287,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprLt(node: PartiqlAst.Expr.Lt, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.LT,
@@ -299,7 +296,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprLte(node: PartiqlAst.Expr.Lte, ctx: Ctx) = visit(node) {
-        Rex.Binary(
+        Plan.rexBinary(
             lhs = convert(node.operands[0]),
             rhs = convert(node.operands[1]),
             op = Rex.Binary.Op.LTE,
@@ -309,7 +306,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
 
     override fun walkExprLike(node: PartiqlAst.Expr.Like, ctx: Ctx) = visit(node) {
         when (val escape = node.escape) {
-            null -> Rex.Call(
+            null -> Plan.rexCall(
                 id = Constants.like,
                 args = args(
                     "value" to node.value,
@@ -317,7 +314,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
                 ),
                 type = StaticType.BOOL,
             )
-            else -> Rex.Call(
+            else -> Plan.rexCall(
                 id = Constants.likeEscape,
                 args = args(
                     "value" to node.value,
@@ -330,7 +327,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprBetween(node: PartiqlAst.Expr.Between, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.between,
             args = args("value" to node.value, "from" to node.from, "to" to node.to),
             type = StaticType.BOOL,
@@ -343,20 +340,20 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
         if (rhs is Rex.Query.Scalar.Subquery) {
             rhs = rhs.query // unpack a scalar subquery coercion
         }
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.inCollection,
             args = listOf(
-                Arg.Value("lhs", lhs),
-                Arg.Value("rhs", rhs),
+                Plan.argValue("lhs", lhs),
+                Plan.argValue("rhs", rhs),
             ),
             type = StaticType.BOOL,
         )
     }
 
     override fun walkExprStruct(node: PartiqlAst.Expr.Struct, ctx: Ctx) = visit(node) {
-        Rex.Tuple(
+        Plan.rexTuple(
             fields = node.fields.map {
-                Field(
+                Plan.field(
                     name = convert(it.first),
                     value = convert(it.second)
                 )
@@ -366,28 +363,28 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprBag(node: PartiqlAst.Expr.Bag, ctx: Ctx) = visit(node) {
-        Rex.Collection.Bag(
+        Plan.rexCollectionBag(
             values = convert(node.values),
             type = StaticType.BAG,
         )
     }
 
     override fun walkExprList(node: PartiqlAst.Expr.List, ctx: Ctx) = visit(node) {
-        Rex.Collection.Array(
+        Plan.rexCollectionArray(
             values = convert(node.values),
             type = StaticType.LIST,
         )
     }
 
     override fun walkExprSexp(node: PartiqlAst.Expr.Sexp, accumulator: Ctx) = visit(node) {
-        Rex.Collection.Array(
+        Plan.rexCollectionArray(
             values = convert(node.values),
             type = StaticType.LIST,
         )
     }
 
     override fun walkExprCall(node: PartiqlAst.Expr.Call, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = node.funcName.text,
             args = args(*node.args.toTypedArray()),
             type = null,
@@ -395,7 +392,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprCallAgg(node: PartiqlAst.Expr.CallAgg, ctx: Ctx) = visit(node) {
-        Rex.Agg(
+        Plan.rexAgg(
             id = node.funcName.text,
             args = listOf(convert(node.arg)),
             modifier = when (node.setq) {
@@ -407,7 +404,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprIsType(node: PartiqlAst.Expr.IsType, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.isType,
             args = args("value" to node.value, "type" to node.type),
             type = StaticType.BOOL,
@@ -415,10 +412,10 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprSimpleCase(node: PartiqlAst.Expr.SimpleCase, ctx: Ctx) = visit(node) {
-        Rex.Switch(
+        Plan.rexSwitch(
             match = convert(node.expr),
             branches = node.cases.pairs.map {
-                Branch(
+                Plan.branch(
                     condition = convert(it.first),
                     value = convert(it.second),
                 )
@@ -429,10 +426,10 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprSearchedCase(node: PartiqlAst.Expr.SearchedCase, ctx: Ctx) = visit(node) {
-        Rex.Switch(
+        Plan.rexSwitch(
             match = null,
             branches = node.cases.pairs.map {
-                Branch(
+                Plan.branch(
                     condition = convert(it.first),
                     value = convert(it.second),
                 )
@@ -470,7 +467,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
                 is PartiqlAst.BagOpType.OuterExcept -> Constants.outerSetExcept
             }
         }
-        Rex.Call(
+        Plan.rexCall(
             id = op,
             args = args("lhs" to node.operands[0], "rhs" to node.operands[1]),
             type = StaticType.BAG,
@@ -478,7 +475,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprCast(node: PartiqlAst.Expr.Cast, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.cast,
             args = args("value" to node.value, "type" to node.asType),
             type = TypeConverter.convert(node.asType),
@@ -486,7 +483,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprCanCast(node: PartiqlAst.Expr.CanCast, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.canCast,
             args = args("value" to node.value, "type" to node.asType),
             type = StaticType.BOOL,
@@ -494,7 +491,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprCanLosslessCast(node: PartiqlAst.Expr.CanLosslessCast, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.canLosslessCast,
             args = args("value" to node.value, "type" to node.asType),
             type = StaticType.BOOL,
@@ -502,7 +499,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprNullIf(node: PartiqlAst.Expr.NullIf, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.nullIf,
             args = args(node.expr1, node.expr2),
             type = StaticType.BOOL,
@@ -510,7 +507,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
     }
 
     override fun walkExprCoalesce(node: PartiqlAst.Expr.Coalesce, ctx: Ctx) = visit(node) {
-        Rex.Call(
+        Plan.rexCall(
             id = Constants.coalesce,
             args = args(node.args),
             type = null,
@@ -519,7 +516,7 @@ internal object RexConverter : PartiqlAst.VisitorFold<RexConverter.Ctx>() {
 
     override fun walkExprSelect(node: PartiqlAst.Expr.Select, ctx: Ctx) = visit(node) {
         when (val query = RelConverter.convert(node)) {
-            is Rex.Query.Collection -> Rex.Query.Scalar.Subquery(query, null)
+            is Rex.Query.Collection -> Plan.rexQueryScalarSubquery(query, null)
             is Rex.Query.Scalar -> query
         }
     }