+++ partiql-plan/src/main/kotlin/org/partiql/plan/Plan.kt
@@ -1,0 +1,6 @@
+package org.partiql.plan
+
+import org.partiql.plan.builder.PlanFactory
+
+// PlanFactory.DEFAULT
+object Plan : PlanFactory()